pyproject.toml
[project]
name = "gpt-shop-viz"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg",
    "pydantic>=2.7",
    "playwright",
    "python-dotenv",
    "alembic",
    "prometheus-client",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "pre-commit",
]
[tool.ruff]
line-length = 100

[tool.ruff.lint]
extend-select = ["I", "B"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
python_version = "3.11"
strict = true
ignore_missing_imports = true
allow_untyped_decorators = true